requirements.txt
openai~=0.11.3
pydub~=0.25.1
PyAudio~=0.2.11
six~=1.16.0
google-cloud-texttospeech~=2.9.0
google-cloud-speech~=2.11.1